refactor: Seperate type, interface
type과 interface를 별도의 모듈로 관리
