--- index.html
@@ -1,14 +1,0 @@
-<!doctype html>
-<html lang="ko">
-<head>
-  <meta charset="UTF-8">
-<!--  <meta name="viewport"-->
-<!--        content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">-->
-<!--  <meta http-equiv="X-UA-Compatible" content="ie=edge">-->
-  <title>Jaydenaver</title>
-  <script defer src="./app.js" type="module" ></script>
-</head>
-<body>
-  <div id="app"></div>
-</body>
-</html>
