--- src/components/header/header.scss
@@ -1,3 +1,0 @@
-.header {
-  color: tomato;
-}
