+++ src/utils/types.ts
@@ -1,26 +1,2 @@
 export type Props = Record<string, string | number | boolean | []>;
 export type State = Record<string, string | number | boolean | []>;
-
-export interface Model {
-  // private _state: State;
-  setState(newState: State): void;
-  get state(): State;
-}
-
-export interface View {
-  // private _template: string;
-  // private _templateElement: HTMLTemplateElement;
-  // private _element: HTMLElement;
-  // private setTemplate(state: State): void;
-  render(state: State): void;
-  get element(): HTMLElement;
-}
-
-export interface Component {
-  // private _model: Model;
-  // private _view: View;
-  // private _element: HTMLElement;
-  setState(state: State): void;
-  get state(): Props;
-  get element(): HTMLElement;
-}
